pyproject.toml
[project]
name = "yt-dynamic-opml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apprise>=1.9.2",
    "fastapi>=0.115.8",
    "google-api-python-client>=2.162.0",
    "oauth2client>=4.1.3",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
